refactored and fixed integration tests, replaced --version flag with --profile-version
